changed the upperbound to decimal...because sometimes 18,446,744,073,709,551,615 just isn't enough.

+++ ClearMeasure.Interview/ClearMeasure.AFA.Tests/UnitTests/CustomDivisorTests.cs
@@ -28,7 +28,7 @@ namespace ClearMeasure.AFA.Tests.UnitTests
             //Arrange
             var suppliedFactorFunctionWasCalled = false;
 
-            Func<ulong, bool> badFactorFunc = num => suppliedFactorFunctionWasCalled = true;
+            Func<decimal, bool> badFactorFunc = num => suppliedFactorFunctionWasCalled = true;
             
             var divisor = new CustomDivisor(3, "", badFactorFunc);
